Provide conformance for messages provided by SwiftProtobuf (#811)
Motivation:

We currently generate conformance to `GRPCProtobufPayload` based on the
input and output types of `rpc`s defined in .proto files passed to the
plugin. This works in some situations, including when the .proto file
for the message in question is not present (such as those provided by
the SwiftProtobuf library; `Google_Protobuf_*`). However this is
problematic if the plugin is invoked on a file-by-file basis where each
file contains the message since the conformance would be erroneously
generated mulutple times.

Modifications:

- Provide conformance to all `Google_Protobuf_*` messages currently
  vendored by `SwiftProtobuf`
- Generaate conformance only based on the `message`s defined in each
  .proto file
- Add a few generation tests

Result:

Multiple conformance generation is avoided in the scenario set out
above.