Implemented file-by-filename and list-services requests  (#1669)
Motivation:

The two requests are necessary for implementing the Reflection Service.

Modifications:

Implemented the ReflectionServiceData struct that stores the useful information from the file descriptor protos received when initiaizing the service and provides the data used for constructing the server response for each type of request. Also imlemented the
functions that create the server responses for the two requests. The requests and the functionality of the ReflectionServiceData struct's methods are tested in the "Tests/GRPCTests/GRPCReflectionServiceTests/GRPCReflectionServiceTests.swift" file.

Result:

The Reflection Service will support the file-by-filename and list-services requests.