Implemented file-containing-extension request for Reflection Service (#1677)
Motivation:

The file-containing-extension request enables users to get the file descriptor protos of
the proto file containing the extension they are looking for and its transitive dependencies.

Modifications:

- Created a new struct (ExtensionDescriptor) to represent the type that is extended and the field number of each extension that
exists inside the protos passed to the Reflection Service.
- Added a <ExtensionDescriptor, FileName> dictionary inside the ReflectionServiceData registry to store the extensions, avoid
duplicates and retrieve nicely the file name of the proto that contains the requested extension. The file name is then used to
get the serialized file descriptor protos of the proto containing the extension and its transitive dependencies.
- Added integration and unit tests.

Result:

The Reflection Service can now be used to get the serialized file descriptor protos of the proto containing the provided
extension and its transitive dependencies.