Update generated code to use service descriptors (#2004)
Motivation:

The generated code refers to services literally using strings. With the
addition of service descriptors, the generated code can now utilize them
to refer to services.

Modifications:

- Update the generated code to use service descriptors.
- For each service in the generated code, extend `ServiceDescriptor` to
include a `static` property that holds the service descriptor for that
service. (I.e. for a service `Foo` in the `bar` package, the service
descriptor for `Foo` can be accessed using `ServiceDescriptor.bar_Foo`.)

Result:

The generated code will utilize and provide easy access to service
descriptors.