Remove unused grpc-swift v1 reflection code (#19)
Motivation:

The reflection code is for v1 and not reference by any target yet. It
needs to be brought up-to-date for v2. Given we need to make a few
changes along the way the diff will be clearer if done from scratch
rather than as a large diff on v1.

Modifications:

- Remove v1 reflection code

Result:

Nothing