Bring HealthService API in line with ReflectionService
Motivation:

The reflection service is a single type called `ReflectionService`. The
Health service is a single type (called `Health`) holding an instance of
the `HealthService` and a type to provide values to the service. This
indirection isn't necessary, the health service can just hold it
instead. This makes the API more like the reflection service and removes
indirection.

Modifications:

- Remove indirection from health service and rename from 'Health' to
  'HealthService'.
- Make the reflection service a struct (it was a class for no good
  reason).

Result:

More consistent API