Add deprecation warnings
Motivation:

gRPC Swift v2 has moved to a new repo, grpc-swift-2 and as a result this
package has a new major version. That's not all that discoverable.

Modifications:

Deprecate commonly used high-level types with a link to a forums post
explaining the move.

Result:

Users are notified about the move