Add more tracing tests (#58)
Motivation:

The OTel tracing interceptors have a few shortcomings
which need fixing. This includes ending the span too
early. This change introduces a number of new tests
which test client and server end-to-end. These tests
currently fail because of the aforementioned issues so
are disabled.

Modifications:

- Add echo service
- Add tracing tests

Result:

Better test coverage