Fix a number of OTel tracing issues (#60)
Motivation:

The OTel tracing interceptors had a number of issues. The most important
was that spans were ended too early in their lifecycle. Others included
not always setting the span status or grpc status code attribute.

Modifications:

- Start and end spans manually
- Replace the hooked async sequence with custom sequnces for server
request messages and client response messages. This allows the atomic to
be dropped in favour of a simple integer stored on the iterator.
- Replace the hooked async writer with a more targeted tracing writer

Result:

Tracing interceptors are more reliable